+++ v2/tests/find_package_build_interface/fpbi-library/src/fpbi.cpp
@@ -1,0 +1,1 @@
+int fpbi_answer() { return 42; }